+++ src/python/config.py
@@ -20,7 +20,7 @@
 # limitations under the License.
 #
 
-version = "1.5.0"
+version = "1.5.1"
 
 verbose = False
 debug = False