+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] 2020-03-21
+
+### Added
+- Added loading tests for all databases
+- Enhanced error reporting (ER #37)
+
+### Changed
+- Added `PROTOCOL=TCPIP` and `AUTHENTICATION=SERVER` to the Db2 connect string.
+- Using `psycopg2.extras.execute_batch()` to speed up data load with Postgres (ER #39)
+
+### Fixed
+- Convert row to tuple on array append (bug #35)
+- Use `UTF-8` connection encoding for Oracle (bug #36)
+
 ## [1.5.0] 2020-03-14
 
 ### Added